Feature/van

+++ src/Services/Carts/Ecommerce.Services.Carts.Api/Models/Dtos/ProductDto.cs
@@ -10,9 +10,9 @@ public class ProductDto
     public decimal UnitPrice { get; set; }
     public decimal DiscountPrice { get; set; }
     public int AvailableStock { get; set; }
-    public double Weight { get; set; }
-    public double Length { get; set; }
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public int Weight { get; set; }
+    public int Length { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
     public string ThumbnailUrl { get; set; } = string.Empty;
 }
