+++ src/Services/Orders/Ecommerce.Services.Orders.Application/Commons/Dtos/Cart/CartItemDto.cs
@@ -14,9 +14,9 @@ public class CartItemDto
     public decimal DiscountPrice { get; set; }
     public long ShopId { get; set; }
     public int AvailableStock { get; set; }
-    public double Weight { get; set; }
-    public double Length { get; set; }
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public int Weight { get; set; }
+    public int Length { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
     public string ThumbnailUrl { get; set; } = string.Empty;
 }
