+++ src/Services/Orders/Ecommerce.Services.Orders.Application/Commons/Dtos/Cart/CheckoutSession.cs
@@ -34,8 +34,8 @@ public class CheckoutSessionItem
     public string? ThumbnailUrl { get; set; }
     public int Quantity { get; set; }
     public decimal UnitPrice { get; set; }
-    public double Weight { get; set; }
-    public double Length { get; set; }
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public int Weight { get; set; }
+    public int Length { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
 }
